pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omniextract-gateway"
version = "0.1.0"
description = "Unified document extraction meta-gateway — engine-agnostic, open source"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["ocr", "llm", "document-extraction", "fastapi", "aws-textract", "azure", "gcp"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core runtime dependencies (API container)
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
# API container extras
api = [
    "litellm>=1.30",
    "redis>=5.0",
]

# OCR Cloud adapter extras
aws = ["boto3>=1.34"]
gcp = ["google-cloud-documentai>=2.0"]
azure = ["azure-ai-documentintelligence>=1.0"]

# Worker / local model extras
worker = [
    "transformers>=4.40",
    "torch>=2.2",
    "Pillow>=10.0",
    "accelerate>=0.29",
]

# Demo client extras
demo = [
    "streamlit>=1.35",
    "httpx>=0.27",
    "Pillow>=10.0",
]

# Development extras
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.9",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "adapters*", "demo*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
